server/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
attrs==23.2.0
cffi==1.16.0
contextlib2==21.6.0
cryptography==42.0.5
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
inflection==0.5.1
injector==0.21.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
pillow==10.2.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
python-magic-bin==0.4.14
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
